Added ability to set a global nabber for styling purposes.
When styling the background of a nav bar there can be a flash of what
during the navigation transitions. This allows you to create a global
nav bar and give it the same color as the other nabbers. It gets
positioned behind everything else.
You set the property $.globalNav = true in a DOM ready function.
You set its background color with the selector: #globaNav.