Refactored $.View virtual stylesheet code.
Added in checks for when a view is imported as ES6 module. In that
case, no element exists yet, so we need to check in the virtual
stylesheet code to prevent element not found error.
Updated $.View mount method to handle creating virtual stylesheet for
imported view. Since at import the view’s anchor element didn’t exist
yet, we need to check during mount for any defined styles and then
scope them to the view.
Added a gulp task to build everything with one task: gulp all.