Fixed some bugs and added some improvements.
1. Removed mount function from router since it wasn’t necessary. No
need to mount imported routes.
2. Fixed view.bindModel so that, if view already has data on `data`
property, this get reset to undefined and then the view is bound to the
model.
3. Fixed view.setData so that, if view is already bound to a model, it
gets unbound and then the views data property gets bound.
4. Updated variables in utilities.js to use ES6 let/const.
5. Fixed bug in model mixing method to update its dataStore properly.
6. Fixed bug in model where a position greater than the model’s array
length was not getting caught properly. Now it will through an error
informing you about the issue. This can be silences with
$.supressErrorMessages set to true.
7. Removed $.formToJSON.js since this functionality is already covered
by $.Form.
8. Refactored $.Form to return JSON object when no validator is
provided.
9. In DOMStack, deprecated `getData()` in favor of `get()`.
10. Fixed bug in array.pluck where unfound properties were resulting in
arrays with empty slots. Now they are ignored.
11. Fixed bug in array.unique where it was returning a copy instead of
removing duplicates directly from the array.